Normalize how email tags are treated
Resolves #143

Previously there was a mismatch between the native macOS app and this
app in how email tags were handled. In the native app if a tag contained
something like an email address it is hidden from the list of tags but
in the Electron app it has been shown in the tag list.

Now the behavior has been matched and in this patch we have normalized
all the places which are checking if a tag contains an email address by
using a new utility `utils/is-email-tag#isEmailTag( tagName )`