Add todo lists in the markdown preview
Formerly this was not working because `marked` didn't render todo lists.
I have switched to a newer markdown parser and added the logic to tie
the render to the underlying note.

Test by viewing the markdown preview and clicking on a todo. The syntax
follows Github Flavored Markdown.

Further, I believe that some of the code highlighting issues may have
been resolved as `highlight.js` now is trying to infer the language on a
block-by-block basis.